+++ changelog.md
@@ -11,6 +11,7 @@ Added:
 * mongoose 支持 mqtt
 * NativeClass::emitSyncFree()
 * spi eth ic driver W5500
+* lcd ic driver ST7789
 
 Fixed: